cookbook: allow for setting --delay duration
The CockroachDB test suite currently makes use of charybdefs-nemesis to
induce artificial delays into the filesystem to simulate disk stalls.
Currently, the delay duration is fixed at 50ms. This precludes testing
of faults that last indefinitely.

Change the format of the `--delay` command to take a duration, in
microseconds, allowing for faults of a longer duration.